requirements.txt
flask==1.1.2
flask_login==0.5.0
flask_migrate==2.4.0
flask_wtf==0.14.3
flask_sqlalchemy==2.*
email_validator==1.1.1
python-decouple==3.3
gunicorn==20.0.4
waitress==1.4.3
python-dotenv==0.14.0
flask-bootstrap==3.3.7.1
flask-mail==0.9.1
psycopg2==2.8.6
flask-jwt-extended==3.24.1
flask-restful==0.3.8
